+++ src/app/page.tsx
@@ -1,103 +1,131 @@
-import Image from "next/image";
+'use client';
 
-export default function Home() {
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import { useRouter } from 'next/navigation';
+import { useEffect, useTransition } from 'react';
+import { useFormState } from 'react-dom';
+import { Toaster, toast } from 'sonner';
+import { Button } from "@/components/ui/button";
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form";
+import { Input } from "@/components/ui/input";
+import { LoginSchema, LoginFormData } from "@/lib/validators";
+import { LoginFormState, loginUser } from "./(auth)/login/actions";
+
+const initialFormState: LoginFormState = {
+  message: undefined,
+  errors: {},
+  success: false,
+};
+
+export default function LoginPage() {
+  const router = useRouter(); // Aunque la redirección principal está en la SA
+  const [isPending, startTransition] = useTransition();
+  const [formState, formAction] = useFormState(loginUser, initialFormState);
+
+  const form = useForm<LoginFormData>({
+    resolver: zodResolver(LoginSchema),
+    defaultValues: {
+      rut: "",
+      password: "",
+    },
+  });
+
+  useEffect(() => {
+    if (formState.success) {
+      // La redirección ya ocurre en la Server Action.
+      console.log("Login exitoso (manejado por Server Action redirect). Mensaje:", formState.message);
+    } else if (formState.message || Object.keys(formState.errors || {}).length > 0) {
+      // Si hay un mensaje general de error
+      if (formState.errors?.general) {
+        toast.error(formState.errors.general);
+      }
+      // Si hay errores específicos de campo, Zod ya los muestra via FormMessage.
+      // Si solo hay un mensaje general y no errores de campo específicos
+      else if (formState.message && !formState.errors?.general && Object.keys(formState.errors || {}).length === 0) {
+         toast.error(formState.message);
+      }
+
+      // Establecer errores específicos de campo en React Hook Form para que FormMessage los muestre
+      if (formState.errors?.rut) {
+        form.setError("rut", { type: "server", message: formState.errors.rut.join(', ') });
+      }
+      if (formState.errors?.password) {
+        form.setError("password", { type: "server", message: formState.errors.password.join(', ') });
+      }
+    }
+  }, [formState, form, router]);
+
+  const onSubmit = (values: LoginFormData) => {
+    startTransition(() => {
+      const formData = new FormData();
+      Object.entries(values).forEach(([key, value]) => {
+        formData.append(key, value as string);
+      });
+      formAction(formData);
+    });
+  };
+
   return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
-        <Image
-          className="dark:invert"
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="list-inside list-decimal text-sm/6 text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
-          <li className="mb-2 tracking-[-.01em]">
-            Get started by editing{" "}
-            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-[family-name:var(--font-geist-mono)] font-semibold">
-              src/app/page.tsx
-            </code>
-            .
-          </li>
-          <li className="tracking-[-.01em]">
-            Save and see your changes instantly.
-          </li>
-        </ol>
+    <> {/* Necesario para Toaster y el div principal */}
+      <Toaster richColors position="top-center" /> {/* Añade el Toaster de Sonner */}
+      <div className="flex items-center justify-center min-h-screen bg-gray-100">
+        <div className="p-8 bg-white shadow-md rounded-lg w-full max-w-md">
+          <h1 className="text-2xl font-bold text-center mb-6">Iniciar Sesión</h1>
+          <Form {...form}>
+            <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-6">
+              <FormField
+                control={form.control}
+                name="rut"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>RUT</FormLabel>
+                    <FormControl>
+                      <Input placeholder="12345678-9" {...field} disabled={isPending} />
+                    </FormControl>
+                    <FormDescription>
+                      Ingresa tu RUT con guion y dígito verificador.
+                    </FormDescription>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
+              <FormField
+                control={form.control}
+                name="password"
+                render={({ field }) => (
+                  <FormItem>
+                    <FormLabel>Contraseña</FormLabel>
+                    <FormControl>
+                      <Input type="password" placeholder="********" {...field} disabled={isPending} />
+                    </FormControl>
+                    <FormMessage />
+                  </FormItem>
+                )}
+              />
 
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:w-auto"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark:invert"
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent font-medium text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 w-full sm:w-auto md:w-[158px]"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
+              {/* El mensaje de error general del servidor ahora se maneja principalmente con toasts */}
+              {formState.message && !formState.success && formState.errors?.general && (
+                <div className="text-sm font-medium text-destructive bg-red-100 p-3 rounded-md mt-2">
+                  {formState.errors.general}
+                </div>
+              )}
+
+              <Button type="submit" className="w-full" disabled={isPending}>
+                {isPending ? "Ingresando..." : "Ingresar"}
+              </Button>
+            </form>
+          </Form>
         </div>
-      </main>
-      <footer className="row-start-3 flex gap-[24px] flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
-    </div>
+      </div>
+    </>
   );
 }
